# btcec: Optimize normalize and NAF, correct normalize, and add tests.